query_params_encode fn moved to types
